+++ com.unity.visualeffectgraph/Editor/Models/Operators/Implementations/TransformVector.cs
@@ -17,10 +17,19 @@ namespace UnityEditor.VFX.Operator
 
         public class OutputProperties
         {
-            public Vector3 tVec = Vector3.zero;
+            public Vector vec;
         }
 
         override public string name { get { return "Transform (Vector)"; } }
+
+        public override void Sanitize(int version)
+        {
+            if (version < 4)
+            {
+                SanitizeHelper.MigrateVector3OutputToSpaceableKeepingLegacyBehavior(this, "Vector");
+            }
+            base.Sanitize(version);
+        }
 
         protected override sealed VFXExpression[] BuildExpression(VFXExpression[] inputExpression)
         {